+++ src/Symfony/Component/JsonEncoder/Mapping/GenericTypePropertyMetadataLoader.php
@@ -18,8 +18,8 @@ use Symfony\Component\TypeInfo\Type\GenericType;
 use Symfony\Component\TypeInfo\Type\IntersectionType;
 use Symfony\Component\TypeInfo\Type\ObjectType;
 use Symfony\Component\TypeInfo\Type\UnionType;
-use Symfony\Component\TypeInfo\TypeContext\TypeContextFactory;
 use Symfony\Component\TypeInfo\Type\WrappingTypeInterface;
+use Symfony\Component\TypeInfo\TypeContext\TypeContextFactory;
 
 /**
  * Enhances properties encoding/decoding metadata based on properties' generic type.