+++ src/Symfony/Component/JsonEncoder/CHANGELOG.md
@@ -5,3 +5,4 @@ CHANGELOG
 ---
 
  * Introduce the component as experimental
+ * Add native PHP lazy ghost support